# openapi/schema.yml
openapi: 3.1.0
info:
  title: Loan Manager API
  version: 1.0.0
  description: Single source of truth for the Loan Manager API. Auth via JWT cookies; FE consumes this spec for types and a thin client.
servers:
  - url: http://localhost:8000/v1
security:
  # Either cookie-based JWT or Bearer token
  - cookieAuth: []
  - bearerAuth: []
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
        correlationId: { type: string }
...
    Collateral:
      type: object
      required: [id, type, value]
      properties:
        id: { type: string }
        loanId: { type: string, nullable: true }
        type: { type: string }
        value: { type: number, minimum: 0 }
        details: { type: string, nullable: true }
    CollateralCreate:
      type: object
      required: [type, value]
      properties:
        id: { type: string }
        type: { type: string }
        value: { type: number, minimum: 0 }
        details: { type: string }
    Vehicle:
      type: object
      required: [id, brand, model, status]
      properties:
        id: { type: string }
        vinOrFrameNumber: { type: string, nullable: true }
        brand: { type: string }
        model: { type: string }
        plate: { type: string, nullable: true }
        color: { type: string, nullable: true }
        purchasePrice: { type: number, minimum: 0, nullable: true }
        msrp: { type: number, minimum: 0, nullable: true }
        status:
          type: string
          enum: [IN_STOCK, ALLOCATED, SOLD]
        linkedLoanId: { type: string, nullable: true }
    VehicleCreate:
      type: object
      required: [brand, model]
      properties:
        id: { type: string }
        vinOrFrameNumber: { type: string }
        brand: { type: string }
        model: { type: string }
        plate: { type: string }
        color: { type: string }
        purchasePrice: { type: number, minimum: 0 }
        msrp: { type: number, minimum: 0 }
...
paths:
...
  /loans/{loanId}/collaterals:
    get:
      summary: List loan collaterals
      parameters: [ { name: loanId, in: path, required: true, schema: { type: string } } ]
      responses:
        '200':
          description: Collaterals
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Collateral' }
    post:
      summary: Create collateral
      parameters: [ { name: loanId, in: path, required: true, schema: { type: string } } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CollateralCreate' }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/Collateral' } } } }
  /loans/{loanId}/collaterals/{collateralId}:
    put:
      summary: Update collateral
      parameters: 
        - { name: loanId, in: path, required: true, schema: { type: string } }
        - { name: collateralId, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CollateralCreate' }
      responses:
        '200': { description: Updated, content: { application/json: { schema: { $ref: '#/components/schemas/Collateral' } } } }
    delete:
      summary: Delete collateral
      parameters: 
        - { name: loanId, in: path, required: true, schema: { type: string } }
        - { name: collateralId, in: path, required: true, schema: { type: string } }
      responses:
        '204': { description: Deleted }
  /vehicle-inventory:
    get:
      summary: List vehicle inventory
      parameters:
        - { name: skip, in: query, schema: { type: integer, minimum: 0, default: 0 } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 1000, default: 100 } }
        - { name: status, in: query, schema: { type: string, enum: [IN_STOCK, ALLOCATED, SOLD] } }
      responses:
        '200':
          description: Vehicles
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Vehicle' }
    post:
      summary: Create vehicle
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/VehicleCreate' }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/Vehicle' } } } }
  /vehicle-inventory/{vehicleId}:
    get:
      summary: Get vehicle
      parameters: [ { name: vehicleId, in: path, required: true, schema: { type: string } } ]
      responses:
        '200': { description: Vehicle, content: { application/json: { schema: { $ref: '#/components/schemas/Vehicle' } } } }
    put:
      summary: Update vehicle
      parameters: [ { name: vehicleId, in: path, required: true, schema: { type: string } } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/VehicleCreate' }
      responses:
        '200': { description: Updated, content: { application/json: { schema: { $ref: '#/components/schemas/Vehicle' } } } }
    delete:
      summary: Delete vehicle
      parameters: [ { name: vehicleId, in: path, required: true, schema: { type: string } } ]
      responses:
        '204': { description: Deleted }
  /vehicle-inventory/{vehicleId}/allocate:
    post:
      summary: Allocate vehicle to loan
      parameters: 
        - { name: vehicleId, in: path, required: true, schema: { type: string } }
        - { name: loanId, in: query, required: true, schema: { type: string } }
      responses:
        '200': { description: Allocated, content: { application/json: { schema: { $ref: '#/components/schemas/Vehicle' } } } }
        '409': { description: Conflict, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /vehicle-inventory/{vehicleId}/release:
    post:
      summary: Release vehicle from allocation
      parameters: [ { name: vehicleId, in: path, required: true, schema: { type: string } } ]
      responses:
        '200': { description: Released, content: { application/json: { schema: { $ref: '#/components/schemas/Vehicle' } } } }
...